+++ RecipaediaEXLoader.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Xml.Linq;
 using Game;
+using RecipaediaEX.Overlay;
 using RecipaediaEX.Search;
 using RecipaediaEX.UI;
 
@@ -15,6 +16,8 @@ namespace RecipaediaEX
             ModsManager.RegisterHook("OnLoadingFinished", this);
             ModsManager.RegisterHook("BlocksInitalized", this);
             ModsManager.RegisterHook("CraftingRecipesManagerInitialized", this);
+            ModsManager.RegisterHook("GuiUpdate", this);
+            ModsManager.RegisterHook("OnModalPanelWidgetSet", this);
         }
 
         public override void OnLoadingFinished(List<Action> actions)
@@ -34,6 +37,20 @@ namespace RecipaediaEX
 
         public override void BlocksInitalized() {
             RecipesCrafterManager.Initialize();
+        }
+
+        public override void GuiUpdate(ComponentGui componentGui) {
+            base.GuiUpdate(componentGui);
+            RecipaediaOverlayInput.HandleRecipaediaKey(componentGui);
+        }
+
+        public override void OnModalPanelWidgetSet(ComponentGui gui, Widget oldWidget, Widget newWidget) {
+            base.OnModalPanelWidgetSet(gui, oldWidget, newWidget);
+            if (newWidget is CraftingTableWidget vanilla && vanilla is not RecipaediaCraftingTableWidget) {
+                gui.ModalPanelWidget = new RecipaediaCraftingTableWidget(
+                    gui.m_componentPlayer.ComponentMiner.Inventory,
+                    vanilla.m_componentCraftingTable);
+            }
         }
 
         public static XElement RequestScreenFile(string screenName) => ContentManager.Get<XElement>($"RecipaediaEX/Screens/{screenName}");